Move sprite eval back to run_cycle.
I've decided that everything that needs to happen on a specific scanline/cycle pair should be done in run_cycle, while everything that needs to be done every pixel (rendering, mixing, sprite0-testing) should be done in their own loops. This is easier to manage, especially given the number of different specific-pixel events that need to be done by the background evaluation.